web_why_lambda/challenge/backend/requirements.txt
Flask
python-dotenv
selenium
flask-session